Add authorization

+++ backend/routes/posts.js
@@ -36,6 +36,7 @@ router.post(
       title: req.body.title,
       content: req.body.content,
       imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then((createdPost) => {
       res.status(201).json({
@@ -65,12 +66,21 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    Post.updateOne({ _id: req.params.id }, post).then((result) => {
-      res.status(200).json({
-        message: "success",
-      });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then((result) => {
+      console.log(result)
+      if (result.modifiedCount > 0) {
+        res.status(200).json({
+          message: "success",
+        });
+      }
+      else {
+        res.status(401).json({
+          message: "Not Authorized",
+        });
+      }
     });
   }
 );
@@ -108,8 +118,12 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
+    if (result.modifiedCount > 0) {
+      res.status(200).json({ message: "Post deleted!" });
+    } else {
+      res.status(200).json({ message: "Not Authorized" });
+    }
   });
 });
